requirements.txt
exa-py
sentence-transformers
faiss-cpu
streamlit
python-dotenv
numpy
